From Michael Osias (IBM)
This patch makes some enhancements to the llRemoteData functions.  The
module is now a shared module, and allows remote data channels to be
created among multiple regions in the same sim.  The port is controlled
from the remoteDataPort property under the [Network] section in
OpenSim.ini.  If this setting is not present or = 0, the module is
disabled and no port is opened.  llRemoteData commands have not effect
when module is disabled.