From: Richard Alimi <ralimi@us.ibm.com>
The following patch implements retrieving prims in Xml2 format via the REST
interface.  For example:
  http://localhost:9000/admin/regions/<region-uuid>/prims/

It also allows an additional parameter which specifies a bounding box.  If
this parameter is specified, only prims within the bounding box are
retrieved.  For example:

http://localhost:9000/admin/regions/8cd759b4-e077-489d-9a34-e1ff70ef65dd/prims/0,0,0,128,128,128
will retrieve only the prims whose positions are in the bounding box with
corners (0,0,0) and (128,128,128).