From: chris yeoh <yeohc@au1.ibm.com>
If you use load-oar to transfer region data from one sim to another
then currently inventory items can be left with unknown owner
permission which results in them being no-mod/no-copy for
everyone. The attached patch fixes things up so if the owner uuid does
not exist on the destination system then it assigns ownership (and the
creator for completeness) to the master avatar id. This will make it
much more practical to share copies of regions :-)