[patching previous patch and also taking the chance of fixing the previous commit message]
This patch reimplements the Draw method in the VectorRenderModule which is
used to create dynamic textures.

The previous version was limited to creating square dynamic textures, it
also didnt allow for dynamically loading an image containing transparency
except at 256x256.

The extraParams string in such functions as osSetDynamicTextureData can
now be passed a comma seperated string of name value pairs which set the
width,height and alpha value of dynamic textures. e.g.
"height:512,width:2048,alpha:255"

Backward compatibility is still preserved so passing the old params of
either a string integer "256" "512" will still work in the same fashion as
will passing "setAlpha" on its own