From: Christopher Yeoh <yeohc@au1.ibm.com>
This patch fixes a bug where if a script in a child prim has taken
control of an avatar when they sit, although permission for camera
control is revoked when they stand, free camera control is not
restored. Currently it is only restored if the script is in the root
prim (though its not clear to me where this happens!).