Adding AssetInventory InventoryArchive plugin
This plugin exposes an HTTP handler on the AssetInventoryServer which
serves a gzipped tar file containing the contents of a user's inventory.
The assets referenced by the inventory are not yet archived. At the
moment only export functionality is implemented, restore functionality
is missing.

prebuild.xml had to be shuffled around a bit in order for the plugin to
build, as it has a dependency on OpenSim.Region.CoreModules.

Also, close a MemoryStream in a few places.