From: Rick Alther <alther@us.ibm.com>
This patch fixes a problem in the RestXmlWriter where an exception may be
thrown when ToString() is called.  ToString() is closing the last tag prior
to returning the string object, which it shouldn't do.  People should be
closing the XML tags and document properly and not rely on the ToString()
side effect to do it for them.  This is most noticeable if you hit
the /regions/ and /regioninfo/ REST calls.  If you are running r9579 or
later, these calls will result in an error.