Enable the console show version command and the viewer about command, to show the last git commit hash together with the conmit date and time. The data is retrieved form a file bin/.version This file can be generated automatically using the post commit script by adding the following to the script:
git log -n 1 --pretty="format:%h: %ci" > bin/.version

This command can also be run manually to create the bin/.version file.

This command genrates a short form of the commit hash and a date and time of the commit in ISO8601 format.
If a full commit hash is required then change %h to %H

The logic that is used to extract the deprecated svn revision is still included.
It will be removed at a future date