Update MySql.Data.dll to version 6.3.6.
This version fixes a bug so that MySqlCommand.CommandTimeout = 0 now correctly sets an infinite command timeout.  This might help with failures on long-running migrations.
The mysql.data.dll v2 is used from the mysql package (as opposed to v4).  The v4 version (probably standing for NET 4.0) does not work under mono 2.6.7, which is still commonly used.